Create new lifecycle exception module
