+++ db/adapters/sqlite/run_adapter.py
@@ -7,7 +7,7 @@ from typing import Optional
 from db.adapters.base import RunDatabaseAdapter
 from db.adapters.sqlite.schema_utils import required_column_names
 from db.adapters.sqlite.sqlite import get_connection, validate_required_fields
-from db.exceptions import DuplicateTurnMetadataError, RunNotFoundError
+from simulation.core.exceptions import DuplicateTurnMetadataError, RunNotFoundError
 from db.schema import runs
 from simulation.core.models.actions import TurnAction
 from simulation.core.models.runs import Run, RunStatus
