+++ simulation/core/query_service.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from db.exceptions import RunNotFoundError
+from simulation.core.exceptions import RunNotFoundError
 from db.repositories.feed_post_repository import FeedPostRepository
 from db.repositories.generated_feed_repository import GeneratedFeedRepository
 from db.repositories.run_repository import RunRepository
