+++ tests/db/adapters/sqlite/test_run_adapter.py
@@ -354,7 +354,7 @@ class TestSQLiteRunAdapterWriteTurnMetadata:
     ):
         """Test that write_turn_metadata raises DuplicateTurnMetadataError when metadata already exists."""
         # Arrange
-        from db.exceptions import DuplicateTurnMetadataError
+        from simulation.core.exceptions import DuplicateTurnMetadataError
 
         run_id = default_test_data["run_id"]
         turn_number = default_test_data["turn_number"]
@@ -496,7 +496,7 @@ class TestSQLiteRunAdapterWriteTurnMetadata:
         but the INSERT fails due to a PRIMARY KEY constraint violation.
         """
         # Arrange
-        from db.exceptions import DuplicateTurnMetadataError
+        from simulation.core.exceptions import DuplicateTurnMetadataError
 
         run_id = default_test_data["run_id"]
         turn_number = default_test_data["turn_number"]
@@ -537,7 +537,7 @@ class TestSQLiteRunAdapterWriteTurnMetadata:
     ):
         """Test that commit is not called when IntegrityError is raised."""
         # Arrange
-        from db.exceptions import DuplicateTurnMetadataError
+        from simulation.core.exceptions import DuplicateTurnMetadataError
 
         run_id = default_test_data["run_id"]
         turn_number = default_test_data["turn_number"]
